--- MySelect/views/Cell.js
@@ -1,8 +1,0 @@
-const Cell = props => {
-  if (!props.data) {
-    return null;
-  }
-  return props.field.options.find(option => option.value === props.data).label;
-};
-
-export default Cell;
