--- Stars/Implementation.js
@@ -1,13 +1,0 @@
-const { Integer } = require('@keystonejs/fields');
-
-class Stars extends Integer.implementation {
-  extendAdminMeta(meta) {
-    return { ...meta, starCount: this.config.starCount || 5 };
-  }
-}
-
-module.exports = {
-  Stars,
-  MongoIntegerInterface: Integer.adapters.mongoose,
-  KnexIntegerInterface: Integer.adapters.knex,
-};
