--- Stars/views/star-full.js
@@ -1,17 +1,0 @@
-import React from "react";
-export default ({ onClick }) => (
-  <svg
-    onClick={onClick}
-    width="22"
-    height="21"
-    viewBox="0 0 44 42"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      stroke="#ED5910"
-      strokeWidth="2"
-      fill="#F8E71C"
-      d="M22 30.972L10.244 39.18l4.175-13.717-11.44-8.643 14.335-.27L22 3l4.686 13.55 14.335.27-11.44 8.643 4.175 13.717z"
-    />
-  </svg>
-);
