--- lists/Image.js
@@ -1,19 +1,0 @@
-const { File} = require('@keystonejs/fields');
-
-
-
-module.exports = (keystone, {imageFileAdapter}) => {
-	console.log('Image');
-
-	keystone.createList('Image', {
-		schemaDoc: 'Image for Article',
-		labelResolver: item => item.file.originalFilename,
-
-		fields: {
-			file: { type: File, adapter: imageFileAdapter }
-
-
-		}
-
-	});
-};
