--- lists/index.js
@@ -1,3 +1,0 @@
-const requireDirectory = require('require-directory');
-const lists = requireDirectory(module);
-module.exports = (keystone, options) => Object.entries(lists).forEach(([key,list]) => list(keystone, options));
