+++ package.json
@@ -11,7 +11,7 @@
     "node": ">=10.0.0"
   },
   "scripts": {
-    "dev": "cross-env NODE_ENV=development DISABLE_LOGGING=true keystone dev --port=4545",
+    "dev": "tsc && cross-env NODE_ENV=development DISABLE_LOGGING=true keystone --entry=tsout/index.js dev --port=4545",
     "build": "cross-env NODE_ENV=production keystone build",
     "start": "cross-env NODE_ENV=production keystone start"
   },
@@ -60,7 +60,9 @@
     "react-apollo": "^3.1.3",
     "react-dom": "^16.12.0",
     "require-directory": "^2.1.1",
-    "sanitize-html": "^1.22.0"
+    "sanitize-html": "^1.22.0",
+    "ts-node": "^8.6.2",
+    "typescript": "^3.8.3"
   },
   "devDependencies": {
     "babel-cli": "^6.26.0",
