+++ tsout/util/emptyObject.js
