+++ util/emptyObject.ts
@@ -1,0 +1,7 @@
+export default obj => {
+    for(const key in obj) {
+        if(obj.hasOwnProperty(key))
+            return false;
+    }
+    return true;
+}
